Move common args to own file
